Prepare for function-providing method
So far, this works:

    python3
    >>> from rpn import rpn
    >>> rpn(["1", "3", "/"])
    0.33333333

and also, this works:

    ./rpn.py 1 3 /
    0.333333

Notice that I now have the .py on the filename.  I think -- but need to
check on this -- that this is required for the 'from' method to work.
But that's just ugly and I think I must be mistaken.

NOTE: wwe need python3 in the above because we are using some python3
formatting methods.  (I do not see any real benefit to this new "f''"
method though, and may change that to see if things work in python2.)